+++ tests/testthat/test-analyze.aov.R
@@ -2,10 +2,9 @@ context("analyze.aov")
 
 test_that("If it works.", {
   library(psycho)
-  library(psycho)
 
   df <- psycho::affective
-  x <- aov(df$Tolerating ~ df$Sex)
+  x <- aov(df$Tolerating ~ df$Salary)
 
   rez <- psycho::analyze(x)
   testthat::expect_equal(nrow(summary(rez)), 2)